🎨 Fix token handling - #4
