+++ minitify/static/js/player.js
@@ -13,7 +13,7 @@ const pathRefreshToken = path.join(tokensPath, 'setting/client.json');
 
 const spotifyApi = new SpotifyWebApi({});
 
-const tokens = JSON.parse(fs.readFileSync(pathAccessToken));
+let tokens = JSON.parse(fs.readFileSync(pathAccessToken));
 let myAccessToken = tokens.myAccessToken;
 const myRefreshToken = tokens.myRefreshToken;
 
@@ -82,6 +82,7 @@ async function refreshAccessToken(){
             };
             let masterData = JSON.stringify(myTokens, null, ' ');
             fs.writeFileSync(pathAccessToken, masterData);
+            ipcRenderer.send("AccessTokenfromPlayerToMain", true);
         },
         err => {
             console.log('Could not refresh access token', err);
@@ -203,86 +204,8 @@ async function skipToBack(){
     }
 }
 
-let favorite = null
-
-function checkSavedTracks(id){
-    spotifyApi.containsMySavedTracks([id])
-    .then(function(data) {
-        const trackIsInYourMusic = data.body[0];
-
-        if (trackIsInYourMusic) {
-            const heart = document.getElementById('heart');
-            heart.setAttribute("style", "fill: red;");
-            console.log('Track was found in the user\'s Your Music library');
-            favorite = true
-        } else {
-            heart.setAttribute("style", "fill: var(--color-heart);");
-            console.log('Track was not found.');
-            favorite = false
-        }
-    }, function(err) {
-        console.log('Something went wrong!', err);
-    });
-}
-
-exports.addOrRemove = addOrRemove
-function addOrRemove(){
-    const heart = document.getElementById('heart');
-    if(favorite){
-        removeFromSaved(id)
-        console.log("remove!!!")
-        favorite = false
-        heart.setAttribute("style", "fill: var(--color-heart);");
-    } else {
-        addToSaved(id)
-        console.log("add")
-        favorite = true
-        heart.setAttribute("style", "fill: red;");
-    }
-}
-
-async function addToSaved(id){
-    spotifyApi.addToMySavedTracks([id])
-    .then(function(data) {
-        console.log('Added track!');
-    }, function(err) {
-        console.log('Something went wrong!', err);
-    });
-}
-
-async function removeFromSaved(id){
-    spotifyApi.removeFromMySavedTracks([id])
-    .then(function(data) {
-        console.log('Removed!');
-    }, function(err) {
-        console.log('Something went wrong!', err);
-    });
-};
-
-
 /* ---------- 🥬 communication ---------- */
 
 function sendfromPlayerToMain(data){
     ipcRenderer.send('fromPlayerToMain', data);
-}
-
-let name = null;
-let id = null;
-let artists = null;
-
-exports.checkFromControllerToMain = checkFromControllerToMain
-function checkFromControllerToMain(){
-    ipcRenderer.send("fromControllerToMain", id);
-}
-
-exports.getfromMainToController = getfromMainToController
-function getfromMainToController(){
-    ipcRenderer.on("fromMainToController", (event, args) => {
-        [name, id, artists] = args;
-        const trackName = document.getElementById('trackName');
-        trackName.textContent = `${name}`;
-        const artistsName = document.getElementById('artistsName');
-        artistsName.textContent = artists;
-        checkSavedTracks(id);
-    });
 }
